minor changes + merge session and session_key mod
